Update for actual allocation of derived type
There was a flaw in the original implementation.

The address of the derived type was clear in the fortran
side (that of the actual target declared in the program),
but in the C side, the address was whatever random data was
in elsi_handle.

This has been corrected now. The initialization routine now
actually allocates space for the type, and returns the address.
In the C side, this means that one has to pass the pointer by
reference (i.e., pass a pointer to it...)